requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
pydantic==2.8.2
networkx==3.3
langchain==0.2.14
langgraph==0.2.24
faiss-cpu==1.8.0.post1
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
streamlit==1.36.0
sentence-transformers==3.0.1
langchain-openai==0.1.22
langchain-community==0.2.12
pytest==8.3.2
prometheus-client==0.20.0
python-dotenv==1.0.1
